Refactored to separate entities from logic models
