+++ QRCodeAuth-Web/QRCodeAuth-Web/Models/Logic/Account.cs
@@ -4,13 +4,13 @@ using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Web;
 
-namespace QRCodeAuth_Web.Models
+namespace QRCodeAuth_Web.Models.Logic
 {
 	public class Account
 	{
 		[Key]
 		public string ID { get; set; }
-		public User Owner { get; set; }
+		public virtual User Owner { get; set; }
 		public string Type { get; set; }
 		public string DepartmentID { get; set; }
 		public string DepartmentName { get; set; }
@@ -18,7 +18,7 @@ namespace QRCodeAuth_Web.Models
 		public bool IsCredentialAuthority { get; set; }
 		public bool IsAttendanceManager { get; set; }
 		public bool IsInformationCollector { get; set; }
-		public List<Event> events { get; set; }
+		public virtual List<Event> events { get; set; }
 
 	}
 }
