+++ QRCodeAuth-Web/QRCodeAuth-Web/Models/Logic/EventType.cs
@@ -3,11 +3,12 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Web;
 
-namespace QRCodeAuth_Web.Models.Domain
+namespace QRCodeAuth_Web.Models.Logic
 {
-	public enum AccountType
+	public enum EventType
 	{
-		Web,
-		Mobile
+		Campus,
+		Lecture,
+		Meeting
 	}
 }
