+++ QRCodeAuth-Web/QRCodeAuth-Web/Models/Logic/User.cs
@@ -4,7 +4,7 @@ using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Web;
 
-namespace QRCodeAuth_Web.Models
+namespace QRCodeAuth_Web.Models.Logic
 {
 	public class User
 	{
@@ -14,7 +14,7 @@ namespace QRCodeAuth_Web.Models
 		public string LastName { get; set; }
 		public string FirstName { get; set; }
 		public UserType Group { get; set; }
-		public List<Account> Accounts { get; set; }
+		public virtual List<Account> Accounts { get; set; }
 
 		public User()
 		{
