+++ QRCodeAuth-Web/QRCodeAuth-Web/Models/Logic/UserType.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Web;
 
-namespace QRCodeAuth_Web.Models
+namespace QRCodeAuth_Web.Models.Logic
 {
 	public enum UserType
 	{
