+++ QRCodeAuth-Web/QRCodeAuth-Web/packages.config
@@ -4,7 +4,7 @@
   <package id="AspNet.ScriptManager.bootstrap" version="3.3.7" targetFramework="net461" />
   <package id="AspNet.ScriptManager.jQuery" version="3.3.1" targetFramework="net461" />
   <package id="bootstrap" version="3.3.7" targetFramework="net461" />
-  <package id="EntityFramework" version="6.3.0" targetFramework="net461" />
+  <package id="EntityFramework" version="6.2.0" targetFramework="net461" />
   <package id="jQuery" version="3.3.1" targetFramework="net461" />
   <package id="Microsoft.AspNet.FriendlyUrls" version="1.0.2" targetFramework="net461" />
   <package id="Microsoft.AspNet.FriendlyUrls.Core" version="1.0.2" targetFramework="net461" />
